bot/requirements.txt
aiohttp>=3.3.0,<3.5.0
git+https://github.com/Rapptz/discord.py@rewrite
pyjwt
websockets>=6.0,<7.0
